requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

numpy>=1.24.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0

pypdf>=3.0.0
python-docx>=0.8.0

# --- dev ---
pytest>=7.4.0
httpx>=0.24.0
black>=23.0.0
isort>=5.12.0
